cmd: Simplify message exchange and parsing
Instead of each command module implementing the same exchange routine,
move that to aux.rs and pass command-specific parser routine to that
exchange routine.

This parser routine validates the received response and collects
all interesting fields from the response to vector of string tuples
that the caller can then use to print/return to user.